Very basic grid based movement and pathfinding
-Shift click to move camera
-scroll wheel to pan in and out
-ctrl+L click to move to that square

Almost everything is completely blank. pathfinding movement is incredibly simple as Unity's AI NavMesh didn't seem to work well.

I have no clue on how the roots are supposed to move , should the object leave a smaller one behind??

Map is very small, because of optimizations. Its going to be one of the more difficult problems to solve. Code will be made more effecient but, it might not be enough.